+++ plugins/PrerenderPlugin/index.js
@@ -302,6 +302,17 @@ class PrerenderPlugin {
 				callback();
 			}
 		});
+
+		compiler.hooks.emit.tapAsync('PrerenderPlugin', (compilation, callback) => {
+			// Replace ReactDOM.render to ReactDOM.hydrate from main.js
+			const data = compilation.assets[opts.chunk].source();
+			const regex = /render(?=(?:(?!render|getElementById).)*getElementById\(('|")root)/;
+			const replacedData = data.replace(regex, 'hydrate');
+
+			emitAsset(compilation, opts.chunk, replacedData);
+
+			callback();
+		});
 	}
 }